+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changelog for pymt_heatf
 2.1 (unreleased)
 ----------------
 
-- Rename component from HeatBMI to HeatModelF (#4)
+- Use Meson for build system (#7)
 - Show total temperature in BMI example (#5)
+- Rename component from HeatBMI to HeatModelF (#4)
 
 
 2.0 (2021-08-24)